INS-2401: RemoteProcedure now receives arguments as byte slice
Previously arguments were passed as two-dimensional slice of bytes. The idea was that each argument should be passed as a separate byteslice, so we get two dimensions. But the experience of RemoteProcedure usage showed that even if we have more than one argument, we still pass all arguments together as a serialized byteslice at the index 0, so we don't need a two-dimensional array.