INS-2503: added important addition to HostNetwork API documentation
All RegisterRequestHandler calls should be executed before Start. This limitation is caused by HostNetwork implementation: packet handlers in Start method are accessed by map that is not protected with any synchronization primitive. So we have to guarantee that all writes to this map (via RegisterRequestHandler) should be performed before Start.