INS-2515: account for pulse in hot data waiter
a message can come close to the end of the current pulse, spend some
time figuring out jet and jet tree and then hang waiting for hot data
until next pulse. this happens when waiter for hot data was refreshed
as pulse has been changed while we were figuring out jet and jet tree.

we discussed this and decided that we wouldn't be saving hot data
history for older pulses. we just make hot data waiter PULSE aware and
make it relase waiter immediately if message is from older pulse.