INS-2579: RPCResponse payload is now nullable
If we make RPCResponse.Result field non-nullable then protobuf compiler thinks that it can't be nil, but in our case it can, because RPCPayload is alias for []byte. So protobuf compiler default-initializes it incorrectly and also does not place some important checks.
That's why we make Result nullable (and protobuf makes it a pointer), although it makes little sense in golang, but now protobuf works correctly.